feat: Add timing output. ...
Add how long it takes to generate the imports. Some modules can take a
long time to build, and the plugin output is unfortunately placed
where no other output can happen for a long time, leading to the
mistaken impression that this plugin can add significant time to the
build.